test: share the App shell mock preset across App-level suites
The four new App-level suites each repeated the same ~100-line block of
vi.mock factories for the shell's chrome components, lazy mode features,
and api modules, plus duplicated installBrowserGlobals/createAppTarget
helpers. The duplication existed only because the suites were authored
by parallel agents under a no-shared-files rule; left in place, every
new App-level suite would copy the block again and any shell dependency
change would need four identical edits.

The invariant mocks now live in lib/testing/appShellMocks.ts (imported
before App.svelte so registration precedes module resolution), with the
instant-onReady runAppStartup mock split into
appShellInstantStartupMock.ts because the startup-timeout suite
exercises the real implementation. Each suite keeps only its genuinely
divergent surfaces inline: the @middleman/ui Provider/view mapping and
the AppHeader treatment. Net -419 lines with identical assertions.

App.test.ts is left on its bespoke mocks: its factories carry stateful
chunk-failure counters and pending-capabilities toggles that do not fit
the shared preset; converging it is possible follow-up but not free.

Validation: all five App-level suites pass (17 tests), full frontend+ui
Vitest suite green (2341 tests), frontend typecheck at its pre-existing
34-error baseline with none in touched files.
