+++ frontend/src/App.force-mobile.test.ts
@@ -10,8 +10,15 @@
 // spec's ".focus-layout .focus-list" check maps to the focus stub being
 // inside main.focus-layout (FocusListView's internal .focus-list markup is
 // covered by FocusListView's own tests).
+//
+// Shell-invariant module mocks come from the appShellMocks preset; only the
+// suite-specific surfaces (@middleman/ui view mapping, AppHeader stub) are
+// mocked here.
 import { cleanup, render, screen, waitFor } from "@testing-library/svelte";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vite-plus/test";
+import "./lib/testing/appShellMocks.js";
+import "./lib/testing/appShellInstantStartupMock.js";
+import { createAppTarget, installBrowserGlobals } from "./lib/testing/appShellMocks.js";
 
 vi.mock("@middleman/ui", async () => {
   const Provider = (await import("./lib/testing/AppProviderMock.svelte")).default;
@@ -37,117 +44,6 @@ vi.mock("@middleman/ui", async () => {
 vi.mock("./lib/components/layout/AppHeader.svelte", async () => ({
   default: (await import("./lib/testing/AppViewKindStub.svelte")).default,
 }));
-vi.mock("./lib/components/layout/StatusBar.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-vi.mock("./lib/components/keyboard/Palette.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-vi.mock("./lib/components/keyboard/Cheatsheet.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-vi.mock("./lib/components/repositories/RepoSummaryPage.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-vi.mock("./lib/components/settings/SettingsPage.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-vi.mock("./lib/components/terminal/WorkspaceTerminalView.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-vi.mock("./lib/components/terminal/WorkspaceEmbedShell.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-vi.mock("./lib/components/design-system/DesignSystemPage.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-vi.mock("./lib/features/kata/KataFeature.svelte", async () => ({
-  default: (await import("./lib/features/kata/KataWorkspaceTestStub.svelte")).default,
-}));
-vi.mock("./lib/features/docs/DocsFeature.svelte", async () => ({
-  default: (await import("./lib/testing/AppDocsFeatureMock.svelte")).default,
-}));
-vi.mock("./lib/features/messages/MessagesFeature.svelte", async () => ({
-  default: (await import("./lib/testing/AppMessagesFeatureMock.svelte")).default,
-}));
-vi.mock("./lib/components/FlashBanner.svelte", async () => ({
-  default: (await import("./lib/testing/AppViewStub.svelte")).default,
-}));
-
-// Pin the container observer to "wide" so the desktop control test below is
-// a true desktop shell (jsdom reports clientWidth 0, which would otherwise
-// classify the container as narrow and trigger the responsive focus
-// presentation even without the force flag).
-vi.mock("./lib/stores/container.svelte.js", () => ({
-  initContainerObserver: () => () => {},
-  getContainerSize: () => "wide",
-  isNarrow: () => false,
-}));
-
-vi.mock("./lib/api/kata/daemons.js", () => ({
-  fetchKataDaemons: vi.fn(async () => []),
-}));
-vi.mock("./lib/api/kata/taskClient.js", () => ({
-  createKataTaskAPI: () => ({}),
-}));
-vi.mock("./lib/api/docs/api.js", () => ({
-  createDocsAPI: () => ({}),
-}));
-vi.mock("./lib/api/messages/api.js", () => ({
-  createMessagesAPI: () => ({
-    capabilities: vi.fn(() =>
-      Promise.resolve({
-        configured: true,
-        ok: true,
-        features: {},
-      }),
-    ),
-  }),
-}));
-vi.mock("./lib/api/messages/visibility.js", () => ({
-  shouldShowMessagesMode: () => true,
-}));
-vi.mock("./lib/messages/kataMessageLinker.js", () => ({
-  createMessageIssueLinker: () => ({
-    linkMessage: vi.fn(),
-  }),
-}));
-vi.mock("./lib/utils/appStartup.js", () => ({
-  runAppStartup: ({ onReady }: { onReady: () => void }) => {
-    queueMicrotask(onReady);
-    return vi.fn();
-  },
-}));
-
-function installBrowserGlobals() {
-  // Desktop environment: every media query (including "(pointer: coarse)")
-  // reports false, matching the Playwright spec's 1280x800 viewport.
-  vi.stubGlobal("matchMedia", (query: string) => ({
-    matches: false,
-    media: query,
-    onchange: null,
-    addEventListener: vi.fn(),
-    removeEventListener: vi.fn(),
-    addListener: vi.fn(),
-    removeListener: vi.fn(),
-    dispatchEvent: vi.fn(),
-  }));
-  vi.stubGlobal(
-    "ResizeObserver",
-    class {
-      observe = vi.fn();
-      unobserve = vi.fn();
-      disconnect = vi.fn();
-    },
-  );
-}
-
-function createAppTarget() {
-  const target = document.createElement("div");
-  target.id = "app";
-  document.body.appendChild(target);
-  return target;
-}
 
 async function renderAppAtIssues() {
   const { replaceUrl } = await import("./lib/stores/router.svelte.ts");
@@ -161,6 +57,8 @@ type ForceMobileWindow = Window & { __MIDDLEMAN_FORCE_MOBILE_ROUTES__?: boolean 
 describe("force-mobile routes flag", () => {
   beforeEach(() => {
     localStorage.clear();
+    // Desktop environment: every media query (including "(pointer: coarse)")
+    // reports false, matching the Playwright spec's 1280x800 viewport.
     installBrowserGlobals();
     window.history.replaceState(null, "", "/");
   });
